requirements.txt
asgiref==3.4.1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
cryptography==3.4.7
defusedxml==0.7.1
Django==3.2.5
django-allauth==0.45.0
django-rest-framework==0.1.0
django-sslserver==0.22
django-widget-tweaks==1.4.8
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
idna==3.2
oauthlib==3.1.1
Pillow==8.3.1
psycopg2==2.9.1
pycparser==2.20
PyJWT==2.1.0
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
urllib3==1.26.6
